Backend/requirements.txt
pandas 
tensorflow
matplotlib
seaborn
opencv-python
easyocr
transformers
pydantic
google-genai
python-dotenv
google-generativeai
jellyfish 
rapidfuzz 
dateparser
langchain
langchain-huggingface
langchain-community
langchain-core
sentence-transformers
fastapi
uvicorn
neo4j
